requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
pandas
numpy
pillow
tqdm
scikit-learn
ipykernel
kagglehub
ultralytics
ipywidgets
mlflow
matplotlib
seaborn
nbformat
streamlit
torchmetrics[detection]
gdown